Review for Location.java on 11/30/2018
Platform:
Java Eclipse 2018-09

Overview:
No bugs. Works as it says it does

Notes/Recommendations:
As it is now, when the constructor is called, these variables are initialized to the following by default:

locId: 0
type: null
occupancy: 0
connection: []

I think it's worth discussing whether or not we should add parameters Location.java's constructor so that these variables are initialized when the object is created.  I bring this up because we may have more debugging if we leave these variables uninitialized.